refactor: addition of headers and num_dllist_t directory
