+++ src/calculator.h
@@ -1,0 +1,2 @@
+#pragma once
+enum exit_code {all_fine = 0, commandline_args_error = 1, no_infix = 2, no_division = 3, input_error = 4, other_errors = 5};
