+++ src/num_dllist/num_dllist.c
@@ -1,0 +1,121 @@
+#include "num_dllist.h"
+#include "../calculator.h"
+#include <stdio.h>
+#include <stdbool.h>
+
+static num_dllist_t *dllist_init(uint8_t *exit_code) {
+  num_dllist_t *dllist = (num_dllist_t *)malloc(sizeof(num_dllist_t));
+
+  if (dllist == NULL) {
+    fprintf(stderr, "Cannot initialize double linked list\n");
+    *exit_code = other_errors;
+    return NULL;
+  }
+
+  dllist->first = NULL;
+  dllist->last = NULL;
+  dllist->sign = SIGN_NOT_SET;
+  dllist->size = 0;
+  dllist->capacity = DLLIST_CAPACITY;
+
+  return dllist;
+}
+
+static void dllist_clear(num_dllist_t *dllist, uint8_t *exit_code) {
+  num_node_t *temp;
+
+  if (dllist == NULL) {
+    fprintf(stderr, "Cannot clear double linked list\n");
+    *exit_code = other_errors;
+  }
+
+  while (dllist->first != NULL) {
+    temp = dllist->first;
+    dllist->first = temp->next;
+    free(temp);
+  }
+
+  dllist->first = NULL;
+  dllist->last = NULL;
+  dllist->sign = SIGN_NOT_SET;
+  dllist->size = 0;
+  dllist->capacity = 0;
+
+  free(dllist);
+}
+
+static void dllist_add_start(num_dllist_t *dllist, uint64_t element,
+                             uint8_t *exit_code) {
+  num_node_t *node = (num_node_t *)malloc(sizeof(num_node_t));
+
+  if ((node == NULL) || (dllist == NULL)) {
+    fprintf(stderr, "Cannot add element to the start of double linked list\n");
+    *exit_code = other_errors;
+  }
+
+  if (dllist->size == dllist->capacity) {
+    fprintf(stderr, "Cannot add element to the start of double linked list because it is full\n");
+    *exit_code = other_errors;
+  }
+
+  node->digit = element;
+  node->next = dllist->first;
+  node->prev = NULL;
+
+  if (dllist->first) dllist->first->prev = node;
+  dllist->first = node;
+
+  if (dllist->last == NULL) dllist->last = node;
+
+  ++dllist->size;
+}
+
+static void dllist_add_end(num_dllist_t *dllist, uint64_t element,
+                           uint8_t *exit_code) {
+  num_node_t *node = (num_node_t *)malloc(sizeof(num_node_t));
+
+  if ((node == NULL) || (dllist == NULL)) {
+    fprintf(stderr,
+            "Cannot add element to the end of the double linked list\n");
+    *exit_code = other_errors;
+  }
+
+  if (dllist->size == dllist->capacity) {
+    fprintf(stderr, "Cannot add element to the end of the double linked list because it is full\n");
+    *exit_code = other_errors;
+  }
+
+  node->digit = element;
+  node->next = NULL;
+  node->prev = dllist->last;
+
+  if (dllist->last) dllist->last->next = node;
+  dllist->last = node;
+
+  if (dllist->first == NULL) dllist->first = node;
+
+  ++dllist->size;
+}
+
+static void dllist_print(num_dllist_t *dllist, uint8_t *exit_code) {
+  num_node_t *temp = dllist->first;
+  bool flag = false;
+
+  if (dllist == NULL) {
+    fprintf(stderr, "Cannot print double linked list\n");
+    *exit_code = other_errors;
+  }
+
+  if (dllist->sign == -1) printf("-");
+
+  while (temp) {
+    if (flag) {
+      printf("%09lu", temp->digit);
+    } else {
+      printf("%ld", temp->digit);
+      flag = true;
+    }
+    temp = temp->next;
+  }
+  printf("\n");
+}
